+++ panda_pytest_assertions/assert_object/_objects_asserters/__init__.py
@@ -3,6 +3,7 @@
 import pytest
 
 pytest.register_assert_rewrite(__name__ + '.equality_asserter')
+pytest.register_assert_rewrite(__name__ + '.identity_asserter')
 pytest.register_assert_rewrite(__name__ + '.is_type_asserter')
 pytest.register_assert_rewrite(__name__ + '.mapping_asserter')
 pytest.register_assert_rewrite(__name__ + '.mapping_subset_asserter')
@@ -15,6 +16,7 @@ pytest.register_assert_rewrite(__name__ + '.with_type_asserter')
 
 from .asserter import Asserter
 from .equality_asserter import EqualityAsserter
+from .identity_asserter import IdentityAsserter
 from .is_type_asserter import IsTypeAsserter
 from .mapping_asserter import MappingAsserter
 from .mapping_subset_asserter import MappingSubsetAsserter
@@ -28,6 +30,7 @@ from .with_type_asserter import WithTypeAsserter
 __all__ = [
     'Asserter',
     'EqualityAsserter',
+    'IdentityAsserter',
     'IsTypeAsserter',
     'MappingAsserter',
     'MappingSubsetAsserter',